projects/contracts/pyproject.toml
[tool.poetry]
name = "smart-contracts"
version = "0.1.0"
description = "AlgoCampus – Algorand smart contracts (ARC-4)"
authors = ["AlgoCampus Team"]
readme = "README.md"
packages = [{ include = "smart_contracts" }]

[tool.poetry.dependencies]
python = "^3.12"
algorand-python = "^2.0.0"
algokit-utils = ">=2.2.0,<3"
py-algorand-sdk = "^2.6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
puyapy = "^3.0.0"
algokit-client-generator = "^1.1.0"
pytest = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
